Resolve clippy warnings surfaced by the new -D warnings CI gate
The stricter clippy invocation added to ci.yml (--all-targets -D warnings)
surfaces lints in test code that the previous CI didn't run with errors.

- per_spherical_tile_source_stack/tests.rs: replace an index-based clear
  with a slice .fill(false) call.
- photometric_ransac/tests.rs: drop a redundant `as usize` cast, and
  allow clippy::needless_range_loop at the module level. The synthetic
  fixtures here use indexed math (target_l[t], truth[i], …); converting
  every loop to .iter().enumerate() would obscure the structure of the
  tests rather than improve it.

+++ crates/sfmtool-core/src/photometric_ransac/tests.rs
@@ -8,6 +8,11 @@
 //! `PerSphericalTileSourceStack`. Tests track the validation plan in
 //! `specs/drafts/photometric-subsets-ransac.md` (items 1–9 and 14–18).
 
+// Indexed loops mirror the indexed math in the synthetic fixtures
+// (`target_l[t]`, `truth[i]`, …); converting them to iterator form would
+// obscure the structure of the tests.
+#![allow(clippy::needless_range_loop)]
+
 use super::*;
 
 // ── Synthesis helpers ────────────────────────────────────────────────────
@@ -203,7 +208,7 @@ fn two_cluster_partition_recovers_both_groups() {
     let out = stack.run(&params);
 
     for t in 0..n_tiles {
-        let a = (t * n_sources) as usize;
+        let a = t * n_sources;
         for i in 0..4 {
             assert!(out.primary_mask[a + i], "tile {t}, src {i}: primary");
             assert!(
